requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Vector database support
pgvector==0.2.4  # PostgreSQL vector extension

# AI and ML
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
openai==1.3.7

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1 